+++ src/main/java/com/single/springboard/util/DateUtils.java
@@ -1,0 +1,10 @@
+package com.single.springboard.util;
+
+import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
+
+public class DateUtils {
+    public static String formatDate(LocalDateTime date) {
+        return date.format(DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss"));
+    }
+}
