+++ src/main/java/com/single/springboard/web/dto/comments/CommentsResponse.java
@@ -9,6 +9,7 @@ public record CommentsResponse(
         String author,
         String content,
         Comments parentId,
-        int commentLevel
+        int commentLevel,
+        String createdDate
 ) {
 }
