+++ src/main/resources/templates/index.html
@@ -41,7 +41,7 @@
                 <th>제목</th>
                 <th>작성자</th>
                 <th>조회수</th>
-                <th>최종수정일</th>
+                <th>등록일</th>
             </tr>
         </thead>
         <tbody>
